### 2020-06-11 #2326 @whedon
[PRE REVIEW]: ‘Bioassays’; a new package in R for analyzing cellular assays
**Submitting author:** @anwarbio (<a href="http://orcid.org/0000-0002-1693-1219">ANWAR PALAKKAN</a>)
**Repository:** <a href="https://github.com/anwarbio/bioassays" target ="_blank">https://github.com/anwarbio/bioassays</a>
**Version:** Version 0.1.0
