+++ paper.bib
@@ -1,0 +1,120 @@
+
+@inbook{Sittampalam:2004,
+  author = {Sittampalam, G. S. and Grossman, A. and Brimacombe, K. and Arkin, M. and Auld, D. and   Austin, C. P. and Baell, J. and Bejcek, B. and Caaveiro, J. M. M. and Chung, T. D. Y. and Coussens, N. P. and Dahlin, J. L. and Devanaryan, V. and Foley, T. L. and Glicksman, M. and Hall, M. D. and Haas, J. V. and Hoare, S. R. J. and Inglese, J. and Iversen, P. W. and Kahl, S. D. and Kales, S. C. and Kirshner, S. and Lal-Nag, M. and Li, Z. and McGee, J. and McManus, O. and Riss, T. and Saradjian, P. and Trask, O. J. and Weidner, J. R. and Wildey, M. J. and Xia, M. and Xu, X.},
+  title = {Assay Guidance Manual},
+  DOI = {NBK144065},
+  url = {https://www.ncbi.nlm.nih.gov/pubmed/23805433},
+  year = {2004},
+  type = {Book Section}
+}
+
+
+@article{Burton:2007,
+  author = {Burton, E. and Yakandawala, N. and LoVetri, K. and Madhyastha, M. S.},
+  title = {A microplate spectrofluorometric assay for bacterial biofilms},
+  journal = {J Ind Microbiol Biotechnol},
+  volume = {34},
+  number = {1},
+  pages = {1-4},
+  ISSN = {1367-5435},
+  DOI = {10.1007/s10295-006-0086-3},
+  url = {https://www.ncbi.nlm.nih.gov/pubmed/16463160},
+  year = {2007},
+  type = {Journal Article}
+}
+
+
+@article{Pant:2016,
+  author = {Pant, K. and Bruce, S. and Sly, J. and Klug Laforce, M. and Springer, S. and Cecil, M. and Andrus, E. and Dakoulas, E. and Wagner, V. O. and Hewitt, N. J. and Kulkarni, R.},
+  title = {Bacterial mutagenicity assays: Vehicle and positive control results from the standard Ames assay, the 6- and 24-well miniaturized plate incorporation assays and the Ames II™ assay},
+  journal = {Environ Mol Mutagen},
+  volume = {57},
+  number = {6},
+  pages = {483-96},
+  ISSN = {1098-2280},
+  DOI = {10.1002/em.22014},
+  url = {https://www.ncbi.nlm.nih.gov/pubmed/27198925},
+  year = {2016},
+  type = {Journal Article}
+}
+
+
+@article{Lin:2015,
+  author = {Lin, A. and Salvador, A. and Carter, J. M.},
+  title = {Multiplexed Microsphere Suspension Array-Based Immunoassays},
+  journal = {Methods Mol Biol},
+  volume = {1318},
+  pages = {107-18},
+  ISSN = {1940-6029},
+  DOI = {10.1007/978-1-4939-2742-5_11},
+  url = {https://www.ncbi.nlm.nih.gov/pubmed/26160569},
+  year = {2015},
+  type = {Journal Article}
+}
+
+
+@article{Song:2015,
+  author = {Song, E. and Yu, M. and Wang, Y. and Hu, W. and Cheng, D. and Swihart, M. T. and Song, Y.},
+  title = {Multi-color quantum dot-based fluorescence immunoassay array for simultaneous visual detection of multiple antibiotic residues in milk},
+  journal = {Biosens Bioelectron},
+  volume = {72},
+  pages = {320-5},
+  ISSN = {1873-4235},
+  DOI = {10.1016/j.bios.2015.05.018},
+url = {https://www.ncbi.nlm.nih.gov/pubmed/26002016},
+  year = {2015},
[…]
