+++ R/bioassays.R
@@ -1,0 +1,16 @@
+#' Bioassays: A Package for Analyzing Multi Well Plate Bioassay's.
+#'
+#' The bioassays package provides three categories of important functions for
+#' extracting, formatting, plotting and analyzing.
+#'
+#' @section Extract functions. The function that help to extract information from file name is \code{\link{extract_filename}}
+#'
+#' @section  Format functions. The function to format data are \code{\link{data2plateformat}},\code{\link{plate2df}},\code{\link{matrix96}},\code{\link{plate_metadata}}
+#'
+#' @section  Plot functions. The function to plot is\code{\link{heatplate}}
+#'
+#' @section  Analysis functions. The function to analyse data are\code{\link{reduceblank}},\code{\link{estimate}},\code{\link{dfsummary}},\code{\link{pvalue}}
+#'
+#' @docType package
+#' @name bioassays
+NULL
