+++ R/rawdata384.R
@@ -1,0 +1,9 @@
+#' rawdata of 384 well plate.
+#'
+#' A dataset of spectrophotometer readings (OD).
+#'
+#' @format A data frame with 16 rows and 25 columns: First column shows 'row name'.
+
+#' @source {Spectrophotometer output reading (OD) of 384 well plate.}
+
+"rawdata384"
