+++ R/rawdata24.R
@@ -1,0 +1,9 @@
+#' rawdata of 24 well plate.
+#'
+#' A dataset of spectrophotometer readings (OD).
+#'
+#' @format A data frame with 4 rows and 7 columns: First column shows 'row name'. .
+
+#' @source {Spectrophotometer output reading (OD) of 24 well plate.}
+
+"rawdata24"
