### 2020-06-12 14c10e6
Update paper.md
