### 2020-06-12 #2326 @labarba
The submission does not currently satisfy our Review Criteria, which require the use of an OSI-approved [Software License](https://joss.readthedocs.io/en/latest/review_criteria.html?highlight=license#software-license).
