### 2020-06-12 #2326 @anwarbio
Hi, I have updated the software License as GPL 3.  Hope this will be fine.
Regards,
Anwar
