### 2020-06-18 #2326 @Kevin-Mattheus-Moerman
@anwarbio I cannot see the license yet. I recommend that you add a plain LICENSE file to your main repository folder. This also means that GitHub "discovers" it so it renders a clear license badge e.g. as circled in blue below.
![Screenshot from 2020-06-18 11-18-40](https://user-images.githubusercontent.com/8392709/85008994-a6620680-b155-11ea-9a04-72454ec98ada.png)
Furthermore I recommend referring to and linking to your LICENSE file in your README document.
