### 2020-06-18 #2326 @Kevin-Mattheus-Moerman
@anwarbio this is not a requirement but I recommend that you change the title to:
`Bioassays: a new package in R for analyzing cellular assays`
i.e. to remove the "special characters"  `'` and `;`. My comment is just to improve accuracy of potential citations to your work in the future.
