+++ README.Rmd
@@ -1,34 +1,0 @@
----
-output: github_document
----
-
-<!-- README.md is generated from README.Rmd. Please edit that file -->
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>",
-  fig.path = "man/figures/README-",
-  out.width = "100%"
-)
-```
-
-# bioassays
-
-<!-- badges: start -->
-<!-- badges: end -->
-
-The goal of bioassays is to help in the analysis of multiwell plate experiments.
-
-## Installation
-
-You can install the released version of bioassays from [CRAN](https://CRAN.R-project.org) with:
-
-``` r
-install.packages("bioassays")
-```
-
-## Example
-
-Detail's of various funtions in this package is provided in 'bioassays-vignette'. Examples on how to use this package is provided in 'bioassays-example' in Vignette.
-
