### 2020-06-18 #2326 @anwarbio
Hi,
I have made all the suggested edits. Details are given below.
1) LICENSE file is added to the repository folder.
2) LICENSE file is  linked to your README document
3) Changed the title as Bioassays: a new package in R for analyzing cellular assays

Regards,
Anwar
