### 2020-06-28 #2402 @whedon
[REVIEW]: ‘Bioassays’; a new package in R for analyzing cellular assays
**Submitting author:** @anwarbio (<a href="http://orcid.org/0000-0002-1693-1219">ANWAR PALAKKAN</a>)
**Repository:** <a href="https://github.com/anwarbio/bioassays" target ="_blank">https://github.com/anwarbio/bioassays</a>
**Version:** v1.0.1

- [x ] **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

The vignettes needed proofreading. I have suggested changes and submitted [a pull request](https://github.com/anwarbio/bioassays/pull/4) for the author to review. (I also included a few other minor suggestions in the pull request, which he might find useful.)
