### 2020-07-03 #2402 @arfon
@anwarbio - your paper title needs quoting, i.e. change this: `title: Bioassays: a new package in R for analyzing cellular assays` to this `title: "Bioassays: a new package in R for analyzing cellular assays"` in your `paper.md`.
