### 2020-07-03 3fcb2e7
Update paper.md
