### 2020-07-03 #2402 @anwarbio
@arfon - The above change is made in the `title`  of `paper.md`
