### 2020-07-03 #2402 @anwarbio
Few mistakes were in the `article proof`, and is now corrected in `paper.md`.
Details of the correction are
1)` Line 25` 1st paragraph last line: _high-throughput_ is corrected as _High-throughput_
2) `Line 27` 2nd paragraph 3rd line: _us_ is corrected as _use_
3) `Line 29` 3rd paragraph 1st line: _server_ is corrected as _serve_
4) `Line 39` 7th paragraph 1st line: An extra _convert_ is deleted.
Thanking you,
Anwar
