### 2020-07-04 #2402 @tomsing1
@fboehm I think I waited too long to accept the https://github.com/openjournals/joss-reviews/invitations invite. Would you mind refreshing it so I can edit the checklist? Many thanks!
