### 2020-07-04 #2326 @arfon
> @whedon add @homoneclocom as reviewer

@fboehm - I'm not sure if you know this already, but you'll need to manually add a checklist for the second reviewer in https://github.com/openjournals/joss-reviews/issues/2402 as Whedon doesn't know how to do this yet.
