### 2020-07-04 #2402 @homonecloco
Hi @anwarbio ,
I started the review with the documentation. I am still to read the paper, but I'll give you my comments of what I have reviewed so far, so you can consider them.

1. The API documentation (https://cran.r-project.org/web/packages/bioassays/bioassays.pdf)  has the beginning looking incomplete or misformated. The documentation of the actual functions seems ok, but having so many NA at the beginning of the document is not recommended.
2. The vignette with examples (https://cran.r-project.org/web/packages/bioassays/vignettes/bioassays-examples.html) refers to some files, like  “L DIFF P3 72HRS.csv”. I couldn’t find that file in the repository. I appreciate that the examples may be included as part of the package, but for this particular document, having the actual file used can help the users to format their tables.
3. In some packages, the documentation is in html and in PDF. It would be useful to have the same option for the vignettes of this packages. The PDFs are hard to copy and paste.
