+++ R/bioassays.R
@@ -5,11 +5,11 @@
 #'
 #' @section Extract functions. The function that help to extract information from file name is \code{\link{extract_filename}}
 #'
-#' @section  Format functions. The function to format data are \code{\link{data2plateformat}},\code{\link{plate2df}},\code{\link{matrix96}},\code{\link{plate_metadata}}
+#' @section  Format functions. The function to format data are \code{\link{data2plateformat}}, \code{\link{plate2df}}, \code{\link{matrix96}}, \code{\link{plate_metadata}}
 #'
-#' @section  Plot functions. The function to plot is\code{\link{heatplate}}
+#' @section  Plot functions. The function to plot is \code{\link{heatplate}}
 #'
-#' @section  Analysis functions. The function to analyse data are\code{\link{reduceblank}},\code{\link{estimate}},\code{\link{dfsummary}},\code{\link{pvalue}}
+#' @section  Analysis functions. The function to analyse data are \code{\link{reduceblank}}, \code{\link{estimate}}, \code{\link{dfsummary}}, \code{\link{pvalue}}
 #'
 #' @docType package
 #' @name bioassays
