### 2020-07-05 anwarbio/bioassays#1 @tomsing1
[JOSS REVIEW]: R CMD check warning and note
Hi @anwarbio,

thank you for submitting your manuscript to JOSS. Great work!
I am one of the two [assigned reviewers](https://github.com/openjournals/joss-reviews/issues/2402) and will file issues for any suggestions I have here. Please feel free to respond with question or comments if anything is unclear or you disagree!

## When I ran `R CMD check` on the package's source code, there was one warning because two of the included filenames contain spaces:

```
Found the following files with non-portable file names:
    inst/exdata/L HEPG2 P3 72HRS.csv
    inst/exdata/metafile 384 plate.csv
  These are not fully portable file names.
  See section ‘Package structure’ in the ‘Writing R Extensions’ manual.
```

It would be great to replace the spaces with underscores (also to keep CRAN happy).

## There was also the following `Note`:

```
File
    LICENSE
  is not mentioned in the DESCRIPTION file.
  Non-standard files/directories found at top level:
    ‘data-raw’ ‘figure1.png’ ‘figure2.png’ ‘figure3.png’ ‘figure4.png’
    ‘paper.bib’ ‘paper.md’
```

1. The `DESCRIPTION` currently refers to `GPL 3` as the license. I think CRAN uses `GPL-3` as the official abbreviation, see [here](https://cran.r-project.org/doc/manuals/r-release/R-exts.html#Licensing).
Right now, it seems that [CRAN](https://cran.r-project.org/web/packages/bioassays/index.html) ignores the `3` suffix in your license declaration and expands it to `GPL-2` and `GPL-3`. Hence, I believe the best line for your  `DESCRIPTION` file would be

```
License: GPL-3 + file LICENSE
```

2. The Note also says `Non-standard files/directories found at top level. These files are your actual manuscript. Adding them to the `.buildignore` file will silence this part of the `Note`, e.g. your `.buildignore` file could look like this:

```
^.*\.Rproj$
^\.Rproj\.user$
data-raw
figure1.png
figure2.png
figure3.png
figure4.png
paper.bib
paper.md
```

These two changes to the DESCRIPTION file silence the note.
