### 2020-07-05 6634e5d
Update bioassays-examples.Rmd
