### 2020-07-05 faee88a
Update bioassays-examples.Rmd
