### 2020-07-05 #2402 @anwarbio
Hi **@homonecloco**

Thank you for reviewing the package and your suggestions are really helpful in improving the package. I have made the below changes in github files. Once all are happy with contents, then I can go for an update on CRAN with all changes.

1. The API documentation (https://cran.r-project.org/web/packages/bioassays/bioassays.pdf) has the beginning looking incomplete or misformated. The documentation of the actual functions seems ok, but having so many NA at the beginning of the document is not recommended.
**comment: I have now corrected this**

2. The vignette with examples (https://cran.r-project.org/web/packages/bioassays/vignettes/bioassays-examples.html) refers to some files, like “L DIFF P3 72HRS.csv”. I couldn’t find that file in the repository. I appreciate that the examples may be included as part of the package, but for this particular document, having the actual file used can help the users to format their tables.
**comment: All the files used in the examples are now added in the exdata folder and is now mentioned in 'vignette with examples'**

3. In some packages, the documentation is in html and in PDF. It would be useful to have the same option for the vignettes of this packages. The PDFs are hard to copy and paste.
**comment: The vignettes are in html format. Adding both html and pdf file might be difficult as the images of  'heatplate' appear in different dimensions in pdf and html.  I might need separate rmd files for html and pdf, I doubt if CRAN will allow this.**

Thanking you,
Amwar
