### 2020-07-05 anwarbio/bioassays#1 @anwarbio
Hi @tomsing1 ,
 Thank you for reviewing my first package, and you suggestion will definitely help me to improve this. I have made the below changes in the github files, and once everybody is happy with this I can update on CRAN.

**`RMD Check:`**
1. **I have now changed the file names. They are now without space**

**`Note:`**
1. **I have made changes in description as** `License: GPL-3 + file LICENSE` .
2. **I have created a `.ignorebuilt` file as you suggested. I am not sure if i have done it right, would be kind if  can check**

Thanking you,
Anwar
