### 2020-07-08 #2402 @homonecloco
Hi @anwarbio ,
Thanks for making the updates so fast. Is there a way I can preview them? Specially I would like to see the NAs go away. In any case, I see you uploaded the file, and it is fine if you only have the html documentation. As long it is available in either format, in an easy to copy and paste (so you can test that the commands are working as expected), is fine.
Best,
Ricardo.
