### 2020-07-08 ba14df0
Update bioassays.Rd
