+++ man/bioassays.Rd
@@ -5,22 +5,36 @@
 \alias{bioassays}
 \title{Bioassays: A Package for Analyzing Multi Well Plate Bioassay's.}
 \description{
-The bioassays package provides three categories of important functions for
-extracting, formatting, plotting and analyzing.
+The bioassays package provides three categories of important functions for extracting, formatting, plotting and analyzing.
 }
-\section{Extract functions. The function that help to extract information from file name is \code{\link{extract_filename}}}{
-NA
+\section{Extract functions}{
+
+The function that help to extract information from file name is \cr
+\code{\link{extract_filename}}'
 }
 
-\section{Format functions. The function to format data are \code{\link{data2plateformat}},\code{\link{plate2df}},\code{\link{matrix96}},\code{\link{plate_metadata}}}{
-NA
+\section{Format functions}{
+
+The function to format data are \cr
+\code{\link{data2plateformat}}\cr
+\code{\link{plate2df}}\cr
+\code{\link{matrix96}}\cr
+\code{\link{plate_metadata}}
 }
 
-\section{Plot functions. The function to plot is\code{\link{heatplate}}}{
-NA
+\section{Plot functions}{
+
+The function to plot is\cr
+\code{\link{heatplate}}
 }
 
-\section{Analysis functions. The function to analyse data are\code{\link{reduceblank}},\code{\link{estimate}},\code{\link{dfsummary}},\code{\link{pvalue}}}{
-NA
+\section{Analysis functions}{
+
+The function to analyse data are\cr
+\code{\link{reduceblank}}\cr
+\code{\link{estimate}}\cr
+\code{\link{dfsummary}}\cr
+\code{\link{pvalue}}
 }
+
 
