### 2020-07-08 #2402 @anwarbio
Hi @homonecloco ,
Thank you for the reply. I have made few more edits today to put each function in separate line (you can see them in the github). I do not know how to share the preview file, so I have shared you a screenshot showing the functions arranged neatly.
Thanking you,
Anwar

![Picture1](https://user-images.githubusercontent.com/7861503/86975899-41586a00-c171-11ea-8957-75ea84540b56.jpg)
