### 2020-07-16 3549ec6
Update bioassays-examples.Rmd
