### 2020-07-24 #2402 @fboehm
@tomsing1 and @homonecloco - I see that @tomsing1 opened three issues in the package repository: https://github.com/anwarbio/bioassays

 @anwarbio responded to the three issues - should we consider the issues closed? Or is more discussion needed?

Thanks again!
