### 2020-08-06 #2402 @homonecloco
In general, I like the extensions to ggplot to display the plates. I can see this plots useful by themselves. It is nice to have some functions that force the structure of the data to the input of the plots, but I don’t think it is completely necessary for a user proficient in R.

- There is a contradiction when comparing the software with DRfit. I would remove the comparison, as DRfit is designed to be user friendly. It may not scale well, but it is not designed for programmers. Also, in the last part of the summary, Bioassays claims to be customisable by anyone using R, which is equivalent to knowing Java to extend Drift.

- The core text oversells the usability of the packages. It is indeed consistent with the user experience that you would get in R, but it is not as “user-friendly” as a graphical interfase. I would tone down the language. For example, the phrase "Since even the most naive computer users know how to change the names of files, this provides a very simple way to pass data of this type into the analysis.”, reads condescending.

- In Acknowledgment, the frase "Support from lab members are acknolegment.” Has a spelling error, and it is not clear which lab members are acknloedege, or at least to which research group.

- The contributions.md stats that there is continues integration in Travis, can you make the link more accessible, to be able to monitor if the current state of the code is running?
