### 2020-08-06 373624c
Update paper.md
