### 2020-08-06 31c4c80
Update appveyor.yml
