+++ appveyor.yml
@@ -12,7 +12,7 @@ install:
 cache:
   - C:\RLibrary
 
-# Adapt as necessary starting from here
+
 
 build_script:
   - travis-tool.sh install_deps
