### 2020-08-06 3ccfa4f
Update CONTRIBUTIONS.md
