### 2020-08-06 #2402 @anwarbio
Hi @homonecloco,

Thank you for the comments. I have made correction as suggested. Details are given below.

1.  In general, I like the extensions to ggplot to display the plates. I can see this plots useful by themselves. It is nice to have some functions that force the structure of the data to the input of the plots, but I don’t think it is completely necessary for a user proficient in R.
**Reply: Any multi-well plate data can be structured for the plots (function 'heatplate') by using functions 'data2plateformat', 'plate2df', 'matrix96' alone or in combinations (depending on the input data). I have included this as a new sentence in line 41**

2. There is a contradiction when comparing the software with DRfit. I would remove the comparison, as DRfit is designed to be user friendly. It may not scale well, but it is not designed for programmers. Also, in the last part of the summary, Bioassays claims to be customisable by anyone using R, which is equivalent to knowing Java to extend Drift.
**Reply: In line 29 the sentence  "They are also limited to the options provided by the software, unless a user is willing to write additional JAVA code, a skill not particularly common among cell biologists" is now removed.**

3. The core text oversells the usability of the packages. It is indeed consistent with the user experience that you would get in R, but it is not as “user-friendly” as a graphical interfase. I would tone down the language. For example, the phrase "Since even the most naive computer users know how to change the names of files, this provides a very simple way to pass data of this type into the analysis.”, reads condescending.
**Reply: This sentence (line 37) is now modified as "This function provides a very easy way to pass data of this type into the analysis by simply editing file names."  Another sentence in line 31 (Moreover, it is strongly documented in a manner designed to be easy for even beginners to grasp.) is modified as "Moreover, we have provided examples (bioassays-examples) in the vignettes, which will be very helpful for beginners to grasp how this package can be used."**

4. In Acknowledgment, the frase "Support from lab members are acknolegment.” Has a spelling error, and it is not clear which lab members are acknloedege, or at least to which research group.
**Reply: The spelling mistake in acknowledgment is now corrected. Lab details are also provided (line 50).**

5. The contributions.md stats that there is continues integration in Travis, can you make the link more accessible, to be able to monitor if the current state of the code is running?
**Reply: The link is now updated.**

Thanking you,
Anwar
