### 2020-08-11 #2402 @fboehm
@anwarbio thank you for the updates! @homonecloco and @tomsing1 - Thanks for your work with the reviews. If you find that you can't yet check some of the boxes in the checklist, please continue to suggest improvements. Please let me know if I can help with anything. Thanks again!!
