### 2020-08-12 #2402 @fboehm
@anwarbio - I failed to produce the pdf. It looks like there's a problem with pandoc-citeproc. By chance did you use the text  "@homonecloco" somewhere in your document? If so, you might need to specify that you aren't intending to cite the key "homonecloco" in your text. You might be able to do this by preceding the at sign by two backslashes, but I could be mistaken on that.
