### 2020-08-12 #2402 @arfon
I can fix the issue here folks, looks like Pandoc is getting confused with the reviewer handles and citations. Pls hold...
