### 2020-08-24 #2402 @fboehm
@homonecloco and @tomsing1 - we have the updated pdf now. What are your feelings about it and the package in the current state? As you work through the last of the checklists, please feel free to continue offering suggestions for improvement as needed, and please do let me know how I can assist you. Thanks again!!
