### 2020-09-17 #2402 @fboehm
@tomsing1 and @homonecloco - It seems that there hasn't been much activity on this review over the last couple of weeks. When you have a moment, would you see if you're comfortable with checking more of the boxes from the checklist? If you're not comfortable checking some boxes, we should be sure that @anwarbio is aware of your concerns. Thanks again!
