### 2020-09-19 #2402 @homonecloco
Hi,
I'm sorry for not commenting earlie. I'm mostly satisfied now. The only thing missing from the guideline is a community guideline on how to contribute to the project. A simple statement in the main readme should be enough.

Best,
Ricardo.
