### 2020-09-20 #2402 @anwarbio
Hi,
@homonecloco Thank you very much.  I have now added a paragraph on contribute in readme.

Kind regards,
Anwar
