### 2020-10-02 82f2b92
Corrected typos and spelling mistakes, reflowed the text to make the code easier to read.
