+++ vignettes/bioassays-vignette.Rmd
@@ -21,13 +21,20 @@ library(bioassays)
 
 # Introduction
 
-In a cell culture lab various cellular assays are performed. The package "bioassays" will help to analyse the results of these experiments performed in multiwell plates. The usage of various functions in the "bioassays" package is provided in this article. 
+In a cell culture lab various cellular assays are performed. The package
+"bioassays" will help to analyse the results of these experiments performed in
+multiwell plates. The usage of various functions in the "bioassays" package is
+provided in this article.
 
-The functions in this package can be used to summarise data from any multiwell plate, and by incorporating them in a loop several plates can be analyzed automatically. Two examples are also provided in the article.
+The functions in this package can be used to summarise data from any multiwell
+plate, and by incorporating them in a loop several plates can be analyzed
+automatically. Two examples are also provided in the article.
 
 # prerequisite
 
-The output reading from the instrument (eg.spectrophotometer) should be in a matrix format. An example data (csv format) is shown below. If the data is in .xls/.xlsx format `read_excel` function in 'readxl' package can be used.
+The output reading from the instrument (eg.spectrophotometer) should be in a
+matrix format. An example data file (csv format) is shown below. If the data is
+in .xls/.xlsx format `read_excel` function from the `readxl` package can be used.
 
 
 ```{r rawdata}
@@ -35,7 +42,9 @@ data(rawdata96)
 head(rawdata96)
 ```
 
-A metadata is needed for the whole experiment. "row" and "col" columns are must in the metafile to indicate the location of well. An example is given below.
+A metadata file is needed for the whole experiment. The `row` and `col` columns
+must be present in the metadata file to indicate the location of well. An
+example is given below.
 
 ```{r metadata}
 data(metafile96)
@@ -46,10 +55,16 @@ head(metafile96)
 
 ## 1. Function: extract_filename
 
-`extract_filename` help to extract information from the file name. syntax is `extract_filename(filename,split = " ",end = ".csv", remove = " ", sep="-")`. 
-**filename** is the file name. **split** is the portions at which the name has to be split (default is space " "). **end** is the extension of file name that need to be removed (default is ".csv"). **remove** is the portion from the file name that need to be omitted after splitting (default is space " "). **sep** add a symbol between separate sections, default is "-".  
+`extract_filename` helps to extract information from the file name. syntax is
+`extract_filename(filename,split = " ",end = ".csv", remove = " ", sep="-")`.
+**filename** is the file name. **split** is the portions at which the name has
+to be split (default is space " "). **end** is the extension of file name that
+need to be removed (default is ".csv"). **remove** is the portion from the file
+name that need to be omitted after splitting (default is space " "). **sep** add
+a symbol between separate sections, default is "-".
 
-This function is useful for extracting specific information from file names, like compound name, plate number etc, to provide appropriate analysis.
+This function is useful for extracting specific information from file names,
+like compound name, plate number etc, to provide appropriate analysis.
 
 For e.g.
 
@@ -63,9 +78,15 @@ extract_filename("L HEPG2 P3 72HRS.csv", split=" ",end=".csv",remove="L",sep="")
 
 ## 2. Function: rmodd_summary
 
[…]
