### 2020-10-02 #2402 @tomsing1
Dear @fboehm and @anwarbio, I have now finished my review. (Thank you for your patience.) I suggested a few improvements in [this pull request](https://github.com/anwarbio/bioassays/pulls). I mainly corrected spelling or grammar mistakes in the vignettes and also added the `.Rbuildignore` file. (I also found one last typo in the paper itself.) The manuscript now fulfills all of the requirements on the list. Good to go!
