+++ paper.bib
@@ -1,10 +1,10 @@
 
-@inbook{Sittampalam:2004,
-  author = {Sittampalam, G. S. and Grossman, A. and Brimacombe, K. and Arkin, M. and Auld, D. and   Austin, C. P. and Baell, J. and Bejcek, B. and Caaveiro, J. M. M. and Chung, T. D. Y. and Coussens, N. P. and Dahlin, J. L. and Devanaryan, V. and Foley, T. L. and Glicksman, M. and Hall, M. D. and Haas, J. V. and Hoare, S. R. J. and Inglese, J. and Iversen, P. W. and Kahl, S. D. and Kales, S. C. and Kirshner, S. and Lal-Nag, M. and Li, Z. and McGee, J. and McManus, O. and Riss, T. and Saradjian, P. and Trask, O. J. and Weidner, J. R. and Wildey, M. J. and Xia, M. and Xu, X.},
+@inbook{Markossian:2004,
+  author = {Markossian, S. and Sittampalam, G. S. and Grossman, A. and Brimacombe, K. and Arkin, M. and Auld, D. and   Austin, C. P. and Baell, J. and Bejcek, B. and Caaveiro, J. M. M. and Chung, T. D. Y. and Coussens, N. P. and Dahlin, J. L. and Devanaryan, V. and Foley, T. L. and Glicksman, M. and Hall, M. D. and Haas, J. V. and Hoare, S. R. J. and Inglese, J. and Iversen, P. W. and Kahl, S. D. and Kales, S. C. and Kirshner, S. and Lal-Nag, M. and Li, Z. and McGee, J. and McManus, O. and Riss, T. and Saradjian, P. and Trask, O. J. and Weidner, J. R. and Wildey, M. J. and Xia, M. and Xu, X.},
   title = {Assay Guidance Manual},
-  DOI = {NBK144065},
-  url = {https://www.ncbi.nlm.nih.gov/pubmed/23805433},
+  url = {https://www.ncbi.nlm.nih.gov/books/NBK53196/},
   year = {2004},
+  Publisher = {Eli Lilly & Company and the National Center for Advancing Translational Sciences},
   type = {Book Section}
 }
 
