### 2020-10-06 972d1c3
Update paper.md
