+++ paper.bib
@@ -5,7 +5,6 @@
   url = {https://www.ncbi.nlm.nih.gov/books/NBK53196/},
   year = {2004},
   Publisher = {Eli Lilly & Company and the National Center for Advancing Translational Sciences},
-  type = {Book Section}
 }
 
 
@@ -20,7 +19,6 @@
   DOI = {10.1007/s10295-006-0086-3},
   url = {https://www.ncbi.nlm.nih.gov/pubmed/16463160},
   year = {2007},
-  type = {Journal Article}
 }
 
 
@@ -35,7 +33,6 @@
   DOI = {10.1002/em.22014},
   url = {https://www.ncbi.nlm.nih.gov/pubmed/27198925},
   year = {2016},
-  type = {Journal Article}
 }
 
 
@@ -49,7 +46,6 @@
   DOI = {10.1007/978-1-4939-2742-5_11},
   url = {https://www.ncbi.nlm.nih.gov/pubmed/26160569},
   year = {2015},
-  type = {Journal Article}
 }
 
 
@@ -61,9 +57,8 @@
   pages = {320-5},
   ISSN = {1873-4235},
   DOI = {10.1016/j.bios.2015.05.018},
-url = {https://www.ncbi.nlm.nih.gov/pubmed/26002016},
+  url = {https://www.ncbi.nlm.nih.gov/pubmed/26002016},
   year = {2015},
-  type = {Journal Article}
 }
 
 
@@ -77,7 +72,6 @@ url = {https://www.ncbi.nlm.nih.gov/pubmed/26002016},
   DOI = {10.1038/srep42383},
   url = {https://www.ncbi.nlm.nih.gov/pubmed/28205527},
   year = {2017},
-  type = {Journal Article}
 }
 
 @article{Wang:2012,
@@ -90,8 +84,7 @@ url = {https://www.ncbi.nlm.nih.gov/pubmed/26002016},
   ISSN = {1367-4811},
   DOI = {10.1093/bioinformatics/btr664},
   url = {https://www.ncbi.nlm.nih.gov/pubmed/22135419},
-year = {2012},
-type = {Journal Article}
+  year = {2012},
 }
 
 
@@ -105,8 +98,7 @@ type = {Journal Article}
   ISSN = {1471-2105},
   DOI = {10.1186/s12859-019-2891-5},
   url = {https://www.ncbi.nlm.nih.gov/pubmed/31113359},
-year = {2019},
-type = {Journal Article}
+  year = {2019},
 }
 
 @Manual{R:2014,
