### 2020-10-07 #2402 @anwarbio
@fboehm -  I have fixed this by removing the word "Journal Article".
Also I have double checked all the dois hyperlink. I can confirm you that all of them are working and directing to correct article.
Thanks
