### 2020-10-07 ad4e74b
Update paper.bib
added braces around Ames and Ames II
