+++ paper.bib
@@ -24,7 +24,7 @@
 
 @article{Pant:2016,
   author = {Pant, K. and Bruce, S. and Sly, J. and Klug Laforce, M. and Springer, S. and Cecil, M. and Andrus, E. and Dakoulas, E. and Wagner, V. O. and Hewitt, N. J. and Kulkarni, R.},
-  title = {Bacterial mutagenicity assays: Vehicle and positive control results from the standard Ames assay, the 6- and 24-well miniaturized plate incorporation assays and the Ames II™ assay},
+  title = {Bacterial mutagenicity assays: Vehicle and positive control results from the standard {Ames} assay, the 6- and 24-well miniaturized plate incorporation assays and the {Ames II™} assay},
   journal = {Environ Mol Mutagen},
   volume = {57},
   number = {6},
