### 2020-10-07 70e7db4
Update NEWS.md
