+++ NEWS.md
@@ -1,4 +1,4 @@
-## version 0.1.0
+## version 1.0.0
 
 ---
 
