### 2020-10-09 f35eae4
Update data2plateformat.R
