### 2020-10-09 d4c51fb
Update extract_filename.R
