+++ R/extract_filename.R
@@ -1,5 +1,30 @@
-
-# function to extract details from file name
+#'@name extract_filename
+#'@aliases extract_filename
+#'@title Extract Information From File Name
+#'@description This function split a string (file name) as per the requirement of the user. It is useful to extract  informations like compound name, plate number etc from the file name.
+#'
+#'@usage extract_filename (filename,split = " ",end = ".csv", remove = " ", sep = "-")
+#'
+#'@param filename name of the file (string).
+#'
+#'@param split regular expressions at which filename has to be split to create different sections.
+#'
+#'@param end extension (end portion) of filename that need to be removed.
+#'
+#'@param remove section that need to be omitted after splitting the filename.
+#'
+#'@param sep symbol to be added to separate sections (obtained after splitting) before combining (default is "-").
+#'
+#'@return A character vector. First element is the unsplit 'filename'. Second element is the processed 'filename'.Other elements are different sections after splitting the 'filename'.
+#'
+#'@author A.A Palakkan
+#'
+#'@examples
+#'extract_filename("L-HEPG2_P3_72HRS.csv")
+#'extract_filename("L_HEPG2_P3_72HRS.csv", split="_",end=".csv",remove="L",sep="")
+#'
+#'@keywords character
+#'
 #' @export
 
 extract_filename<-function(filename,split = " ",end = ".csv", remove = " ", sep="-"){
