### 2020-10-09 9d75acf
Update heatplate.R
