### 2020-10-09 e47e9f4
Update rawdata24.R
