+++ R/rawdata24.R
@@ -1,9 +1,16 @@
-#' rawdata of 24 well plate.
+#'@docType data
+#'@name rawdata24
+#'@aliases rawdata24
+#'@title Rawdata of 24 well plate.
 #'
-#' A dataset of spectrophotometer readings (OD).
+#' @format A data frame with 4 rows and 7 columns: First column shows 'row name'.
 #'
-#' @format A data frame with 4 rows and 7 columns: First column shows 'row name'. .
-
 #' @source {Spectrophotometer output reading (OD) of 24 well plate.}
-
-"rawdata24"
+#'
+#'@usage rawdata24
+#'
+#'@description A dataset of spectrophotometer readings (OD).
+#'
+#'@keywords datasets
+#'
+NULL
