+++ R/rawdata384.R
@@ -1,9 +1,14 @@
-#' rawdata of 384 well plate.
-#'
-#' A dataset of spectrophotometer readings (OD).
+#'@docType data
+#'@name rawdata384
+#'@aliases rawdata384
+#'@title Rawdata of 384 well plate.
 #'
 #' @format A data frame with 16 rows and 25 columns: First column shows 'row name'.
-
 #' @source {Spectrophotometer output reading (OD) of 384 well plate.}
-
-"rawdata384"
+#'
+#' @usage rawdata384
+#'
+#' @description A dataset of spectrophotometer readings (OD).
+#'
+#' @keywords datasets
+NULL
