### 2020-10-09 e46a32c
Update rawdata96.R
