+++ R/rawdata96.R
@@ -1,9 +1,14 @@
-#' rawdata of 96 well plate.
-#'
-#' A dataset of spectrophotometer readings (OD).
+#'@docType data
+#'@name rawdata96
+#'@aliases rawdata96
+#'@title Rawdata of 96 well plate.
 #'
 #' @format A data frame with 8 rows and 13 columns: First column shows 'row name'.
-
 #' @source {Spectrophotometer output reading (OD) of 96 well plate.}
-
-"rawdata96"
+#'
+#'@usage rawdata96
+#'
+#'@description A dataset of spectrophotometer readings (OD).
+#'
+#'@keywords datasets
+NULL
