### 2020-10-09 102b408
Update reduceblank.R
