### 2020-10-09 7e05870
Update NEWS.md
