+++ NEWS.md
@@ -1,5 +1,5 @@
 ## version 1.0.0
-
+- OCT 2020
 ---
 
 ### NEWS.md setup
