### 2020-10-10 #2402 @kthyng
Ok I can take over from here! I see that the version and Zenodo archive are taken care of! Paper looks good too. We can move forward.
